requirements.txt
langchain
langgraph
langchain-openai
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
python-dotenv
